Pyflakes: HUGE cleanup
It was quite a mess. Double imports, unused imports, imports that overwrite other imports.

Added a couple of 'TODO' comments, mostly for python's logfile system that gets initialized at the
top level of the module. Don't do that. Do that in the main() where you need it.
And don't set it in the tests.